Multi-page form reporting.
This splits the reporting form up into multiple 'pages', such as
category, extra questions, and photo upload. On mobile, it moves you
back to the map if there are assets to check/pick, or if the road
not found event has fired.

[TfL] One thing to note is that, due to changes in duplicate showing,
the category change event is no longer triggered, which was the only
thing allowing the TfL red route message to be shown when moving from a
point asset category such as Bus stops to a red-route only category. So
we swap the asset/road layers so any asset is deselected before red
route checking. The converse problem, where the red route checking would
remove the asset message if it came after, was hopefully fixed by the
“Only hide relevant responsibility message” commit.