wip
Always set category_groups, even if not in use, so templates can be
simpler.

`mix_in` will be used in places where we want category_groups be a mix
of groups (for those categories that have them) and categories (for
those without a group) that can be output directly with the
subcategories interspersed in the middle rather than all without a group
shown first (which then gets changed anyway if JS to mingle in the
groups). Might make the JS setting it up bit easier or not much
different anyway, but when we come to change select to radios on
reporting, should make it easier to then replace the headings with
buttons, move the subcategories down to then appear underneath instead
when selected.