Add top_banner functionality
Created front.js file, updated index.html and layout.scss files.
Changes made will be recorded in the Pull Request description of the original popravi.to repository. Likewise, they will be recorded in the commit comment of the forked popravi.to repository.
This commit contains also the changes made by the 44b96ec commit which are not as of yet pulled into the original popravi.to repository.
This addresses issue #34.

This issue was worked on together by next GitHub users
@MihajloVSc
@jerkdavi